Refactor query yes/no to utilities.py
